revert moving queue traits to its own file
