+++ libs/traits/src/liquidity_pools.rs
@@ -11,7 +11,7 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use frame_support::dispatch::DispatchResult;
+use frame_support::{dispatch::DispatchResult, weights::Weight};
 use sp_runtime::DispatchError;
 use sp_std::vec::Vec;
 
@@ -92,6 +92,27 @@ pub trait MessageReceiver {
 		origin: Self::Origin,
 		message: Vec<u8>,
 	) -> DispatchResult;
+}
+
+/// The trait required for queueing messages.
+pub trait MessageQueue {
+	/// The message type.
+	type Message;
+
+	/// Submit a message to the queue.
+	fn queue(msg: Self::Message) -> DispatchResult;
+}
+
+/// The trait required for processing dequeued messages.
+pub trait MessageProcessor {
+	/// The message type.
+	type Message;
+
+	/// Process a message.
+	fn process(msg: Self::Message) -> (DispatchResult, Weight);
+
+	/// Max weight that processing a message can take
+	fn max_processing_weight(msg: &Self::Message) -> Weight;
 }
 
 /// The trait required for handling outbound LP messages.
