+++ pallets/liquidity-pools-gateway/src/message_processing.rs
@@ -1,6 +1,5 @@
-use cfg_traits::{
-	liquidity_pools::{InboundMessageHandler, LPMessage, MessageHash, RouterProvider},
-	queue::MessageQueue,
+use cfg_traits::liquidity_pools::{
+	InboundMessageHandler, LPMessage, MessageHash, MessageQueue, RouterProvider,
 };
 use cfg_types::domain_address::{Domain, DomainAddress};
 use frame_support::{
